requirements.txt
Django==1.10.4
dj-database-url==0.4.1
dj-static==0.0.6
gunicorn==19.6.0
Unipath==1.1
python-decouple==3.0
Pillow==3.4.2
Markdown==2.6.7
bleach==1.5.0
whitenoise==3.2.2
psycopg2==2.6.2
django-closuretree==1.1.10
pyx
djangorestframework
djangorestframework-jwt
django-filter

